requirements.txt
psycopg2-binary==2.9.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.15
PyQt5_sip==12.15.0
